feat: refactor types to new file
