+++ src/describe/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2020, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export { SObjectService, SObjectType } from './sobjectService';
+export { SObject } from './sobjectTypes';
